+++ src/deps.ts
@@ -1,4 +1,4 @@
 export {
   default as Semver,
   eq,
-} from "https://deno.land/x/semver@v1.4.0/mod.ts";
+} from "https://deno.land/std@0.173.0/semver/mod.ts";