+++ test/fixture/CustomRelease.ts
@@ -1,8 +1,12 @@
-import { Release, Change } from "../../mod.ts";
+import { Change, Release } from "../../mod.ts";
 import { Semver } from "../../src/deps.ts";
 
 class CustomRelease extends Release {
-  constructor(version?: Semver | string, date?: Date | string, description?: string) {
+  constructor(
+    version?: Semver | string,
+    date?: Date | string,
+    description?: string,
+  ) {
     super(version, date, description);
     this.changes.set("maintenance", []);
   }
@@ -11,6 +15,10 @@ class CustomRelease extends Release {
   }
 }
 
-export default function (version?: Semver | string, date?: Date | string, description?: string) {
+export default function (
+  version?: Semver | string,
+  date?: Date | string,
+  description?: string,
+) {
   return new CustomRelease(version, date, description);
 }